requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
youtube-dl==2021.12.17
tweepy==4.14.0
langchain
langchain-community
faiss-cpu==1.7.4
numpy<2
pandas==2.0.3
vaderSentiment==3.3.2
sentence-transformers>=2.7
huggingface-hub>=0.17
openai==1.3.5
python-dotenv
